requirements.txt
pandas==0.24.1
scikit-learn
sklearn-pandas
numpy

